make CanvasIntegration an interface
